View rule tabpane
